microblogging/requirements.txt
asgiref==3.5.0
Django==4.0.3
django-js-routes==0.2.0
django-webpack-loader==1.4.1
inertia-django==0.3.2
marshmallow==3.15.0
packaging==21.3
pyparsing==3.0.7
sqlparse==0.4.2
